python/car_denting_repair_charges/requirements.txt
python-dotenv==1.0.0
pillow==10.0.0
opencv-python==4.8.0.76
numpy==1.24.3
openai==0.27.8
anthropic==0.7.1
requests==2.31.0
pydantic==2.3.0
pytest==7.4.0
pytest-cov==4.1.0
pytest-mock==3.11.1
mock==5.1.0
python-jose==3.3.0
